Working on collision policies.
Implement method for checking collisions. Only (aisle == 2 tracks)
restriction at the time of this commit, (cross == 3 tracks) restriction remaining.
This fixes a bug we had because of checking collisions of items instead of routes.

Surround correctly the methods and class definition in route.py with blank lines
according to pep 8.

Modify .gitignore. Add aux file to write commits.

Run a test starting with

last_item = {
 "sku" : "ASYQU8", "weight" : 10, "volume" : 40, "stock_label" : "MZ-1-003-024"
}

to test distance func from an item in the center of the map.