Running Version with new collision restrictions.
No big zigzags allowed.

FIX BUG, as we implemented the dummy first item for the first route, we
never entered a new turn.

FIX BUG when writing a new tikzpicture in tex for new turn.

Add drawing functionality in latex.

Add batch file for experiments.

Organize a sample/ folder as an experimentation sandbox.